unit tests and benchmarks for base64decode
